+++ driver/src/orderbook/streamed/mod.rs
@@ -1,5 +1,6 @@
 #![allow(dead_code)]
 
+mod error;
 mod state;
 
 use ethcontract::Address;
